extract popover content into separate components
